Project Rock Paper Scissors, Steps 1-5
Created two functions and logic to run a rock, paper, scissors game.
The computer makes a random guess based on a random number that gets
converted to one of the choices. The user inputs their guess in the
code, since I have not built the prompt or GUI yet.

Part of Odin Project foundations course